split client::request -> client::api::{http, in_memory}
- rename `crate::client::request` into `crate::client::api`
- split the module in three parts:
  - the `ApiClient` trait definition in `api/mod.rs`
  - the in memory impl in `api/in_memory.rs`
  - the HTTP impl in `api/http.rs`
